Feature/lab 2 planet 0002 base interfaces for core structure of the application

+++ src/main/java/core/Product.java
@@ -11,11 +11,11 @@ public class Product {
     private String productName;
 
     /**
-     * Price of the product in polish zł
+     * Price of the product in polish zł.
      */
-    private int price;
+    private double price;
 
-    public Product(String productName, int price) {
+    public Product(String productName, double price) {
         this.productName = productName;
         this.price = price;
     }
@@ -28,11 +28,11 @@ public class Product {
         this.productName = productName;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
